+++ src/rust-crypto/RoomEncryptor.ts
@@ -33,6 +33,7 @@ import { KeyClaimManager } from "./KeyClaimManager";
 import { RoomMember } from "../models/room-member";
 import { HistoryVisibility } from "../@types/partials";
 import { OutgoingRequestsManager } from "./OutgoingRequestsManager";
+import { logDuration } from "../utils";
 
 /**
  * RoomEncryptor: responsible for encrypting messages to a given room
@@ -44,6 +45,9 @@ export class RoomEncryptor {
 
     /** whether the room members have been loaded and tracked for the first time */
     private lazyLoadedMembersResolved = false;
+
+    /** Ensures that there is only one call to shareRoomKeys at a time */
+    private currentShareRoomKeyPromise = Promise.resolve();
 
     /**
      * @param olmMachine - The rust-sdk's OlmMachine
@@ -95,8 +99,10 @@ export class RoomEncryptor {
             (member.membership == "invite" && this.room.shouldEncryptForInvitedMembers())
         ) {
             // make sure we are tracking the deviceList for this user
-            this.olmMachine.updateTrackedUsers([new UserId(member.userId)]).catch((e) => {
-                this.prefixedLogger.error("Unable to update tracked users", e);
+            logDuration(this.prefixedLogger, "updateTrackedUsers", async () => {
+                this.olmMachine.updateTrackedUsers([new UserId(member.userId)]).catch((e) => {
+                    this.prefixedLogger.error("Unable to update tracked users", e);
+                });
             });
         }
 
@@ -113,7 +119,10 @@ export class RoomEncryptor {
      */
     public async prepareForEncryption(globalBlacklistUnverifiedDevices: boolean): Promise<void> {
         const logger = new LogSpan(this.prefixedLogger, "prepareForEncryption");
-        await this.ensureEncryptionSession(logger, globalBlacklistUnverifiedDevices);
+
+        await logDuration(this.prefixedLogger, "prepareForEncryption", async () => {
+            await this.ensureEncryptionSession(logger, globalBlacklistUnverifiedDevices);
+        });
     }
 
     /**
@@ -156,7 +165,10 @@ export class RoomEncryptor {
             // at the end of the sync, but we can't wait for that).
             // XXX future improvement process only KeysQueryRequests for the users that have never been queried.
             logger.debug(`Processing outgoing requests`);
-            await this.outgoingRequestManager.doProcessOutgoingRequests();
+
+            await logDuration(this.prefixedLogger, "doProcessOutgoingRequests", async () => {
+                await this.outgoingRequestManager.doProcessOutgoingRequests();
+            });
         } else {
             // If members are already loaded it's less critical to await on key queries.
             // We might still want to trigger a processOutgoingRequests here.
@@ -174,7 +186,10 @@ export class RoomEncryptor {
         );
 
         const userList = members.map((u) => new UserId(u.userId));
-        await this.keyClaimManager.ensureSessionsForUsers(logger, userList);
+
+        await logDuration(this.prefixedLogger, "ensureSessionsForUsers", async () => {
+            await this.keyClaimManager.ensureSessionsForUsers(logger, userList);
+        });
 
         const rustEncryptionSettings = new EncryptionSettings();
         rustEncryptionSettings.historyVisibility = toRustHistoryVisibility(this.room.getHistoryVisibility());
@@ -198,16 +213,38 @@ export class RoomEncryptor {
         rustEncryptionSettings.onlyAllowTrustedDevices =
             this.room.getBlacklistUnverifiedDevices() ?? globalBlacklistUnverifiedDevices;
 
-        const shareMessages: ToDeviceRequest[] = await this.olmMachine.shareRoomKey(
-            new RoomId(this.room.roomId),
-            userList,
-            rustEncryptionSettings,
-        );
-        if (shareMessages) {
-            for (const m of shareMessages) {
-                await this.outgoingRequestManager.outgoingRequestProcessor.makeOutgoingRequest(m);
+        await logDuration(this.prefixedLogger, "shareRoomKey", async () => {
+            const shareMessages: ToDeviceRequest[] = await this.shareRoomKey(userList, rustEncryptionSettings);
+            if (shareMessages) {
+                for (const m of shareMessages) {
+                    await this.outgoingRequestManager.outgoingRequestProcessor.makeOutgoingRequest(m);
+                }
             }
-        }
+        });
+    }
+
+    /**
+     *  The Rust-SDK requires that we only have one shareRoomKey process in flight at once for a room.
+     *  This method ensures that, by only having one call to shareRoomKey active at once (and making them
+     *  queue up in order).
+     *
+     * @param userList - list of userIDs to share with
+     * @param rustEncryptionSettings - encryption settings to use
+     *
+     * @returns a promise which resolves to the list of ToDeviceRequests to send
+     */
+    private async shareRoomKey(
+        userList: UserId[],
+        rustEncryptionSettings: EncryptionSettings,
+    ): Promise<ToDeviceRequest[]> {
+        const prom = this.currentShareRoomKeyPromise
+            .catch(() => {
+                // any errors in the previous claim will have been reported already, so there is nothing to do here.
+                // we just throw away the error and start anew.
+            })
+            .then(() => this.olmMachine.shareRoomKey(new RoomId(this.room.roomId), userList, rustEncryptionSettings));
+        this.currentShareRoomKeyPromise = prom;
+        return prom;
     }
 
     /**